Update wrong nuget package information
I noticed that the prerequisites nuget packages have wrong (outdated) information associated with them. This creates great confusion for anyone not familiar with the packages history. The "Azure Storage Common" nuget package is no longer in preview, and the "Microsoft Azure CosmosDB Table" nuget package linked to has a dependency is .NET Full framework, and not .net standard as eluded to in the current text. There is however a preview version of a .net standard package which should be referenced as well.